Change all time db to bigint

+++ db/migrate/20210612055652_create_worktimes.rb
@@ -1,10 +1,10 @@
 class CreateWorktimes < ActiveRecord::Migration[6.0]
   def change
     create_table :worktimes do |t|
-      t.integer :start_time
-      t.integer :pause_time
-      t.integer :accumulate_time
-      t.integer :end_time
+      t.bigint :start_time
+      t.bigint :pause_time
+      t.bigint :accumulate_time
+      t.bigint :end_time
       t.references :user, foreign_key: true
 
       t.timestamps
